refactor: move timeUtils
